+++ attributes.go
@@ -43,3 +43,19 @@ const (
 	// string parameters are removed.
 	AttributeRequestReferer = "request.headers.referer"
 )
+
+// Attributes destined for Span Events:
+//
+// To disable the capture of one of these span event attributes, db.statement
+// for example, modify your Config like this:
+//
+//	cfg.SpanEvents.Attributes.Exclude = append(cfg.SpanEvents.Attributes.Exclude,
+//		newrelic.SpanAttributeDBStatement)
+const (
+	SpanAttributeDBStatement  = "db.statement"
+	SpanAttributeDBInstance   = "db.instance"
+	SpanAttributePeerAddress  = "peer.address"
+	SpanAttributePeerHostname = "peer.hostname"
+	SpanAttributeHTTPURL      = "http.url"
+	SpanAttributeHTTPMethod   = "http.method"
+)