+++ internal_span_events_test.go
@@ -1,6 +1,7 @@
 package newrelic
 
 import (
+	"net/http"
 	"testing"
 
 	"github.com/newrelic/go-agent/internal"
@@ -107,4 +108,356 @@ func TestSpanEventsDisabledWithoutDistributedTracing(t *testing.T) {
 	segment.End()
 	txn.End()
 	app.ExpectSpanEvents(t, []internal.WantEvent{})
+}
+
+func TestSpanEventDatastoreExternal(t *testing.T) {
+	// Test that a datastore and external segments creates the correct span
+	// events.
+	replyfn := func(reply *internal.ConnectReply) {
+		reply.AdaptiveSampler = internal.SampleEverything{}
+	}
+	cfgfn := func(cfg *Config) {
+		cfg.DistributedTracer.Enabled = true
+		cfg.CrossApplicationTracer.Enabled = false
+	}
+	app := testApp(replyfn, cfgfn, t)
+	txn := app.StartTransaction("hello", nil, nil)
+	segment := DatastoreSegment{
+		StartTime:          StartSegmentNow(txn),
+		Product:            DatastoreMySQL,
+		Collection:         "mycollection",
+		Operation:          "myoperation",
+		ParameterizedQuery: "myquery",
+		Host:               "myhost",
+		PortPathOrID:       "myport",
+		DatabaseName:       "dbname",
+	}
+	segment.End()
+	req, _ := http.NewRequest("GET", "http://example.com?ignore=me", nil)
+	s := StartExternalSegment(txn, req)
+	s.End()
+	txn.End()
+	app.ExpectSpanEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":          "OtherTransaction/Go/hello",
+				"sampled":       true,
+				"category":      "generic",
+				"nr.entryPoint": true,
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"sampled":   true,
+				"name":      "Datastore/statement/MySQL/mycollection/myoperation",
+				"category":  "datastore",
+				"component": "MySQL",
+				"span.kind": "client",
+			},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"db.statement":  "myquery",
+				"db.instance":   "dbname",
+				"peer.address":  "myhost:myport",
+				"peer.hostname": "myhost",
+			},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"name":      "External/example.com/all",
+				"category":  "http",
+				"component": "http",
+				"span.kind": "client",
+			},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"http.url":    "http://example.com",
+				"http.method": "GET",
+			},
+		},
+	})
+}
+
+func TestSpanEventAttributesDisabled(t *testing.T) {
+	// Test that SpanEvents.Attributes.Enabled correctly disables span
+	// attributes.
+	replyfn := func(reply *internal.ConnectReply) {
+		reply.AdaptiveSampler = internal.SampleEverything{}
+	}
+	cfgfn := func(cfg *Config) {
+		cfg.DistributedTracer.Enabled = true
+		cfg.CrossApplicationTracer.Enabled = false
+		cfg.SpanEvents.Attributes.Enabled = false
+	}
+	app := testApp(replyfn, cfgfn, t)
+	txn := app.StartTransaction("hello", nil, nil)
+	segment := DatastoreSegment{
+		StartTime:          StartSegmentNow(txn),
+		Product:            DatastoreMySQL,
+		Collection:         "mycollection",
+		Operation:          "myoperation",
+		ParameterizedQuery: "myquery",
+		Host:               "myhost",
+		PortPathOrID:       "myport",
+		DatabaseName:       "dbname",
+	}
+	segment.End()
+	req, _ := http.NewRequest("GET", "http://example.com?ignore=me", nil)
+	s := StartExternalSegment(txn, req)
+	s.End()
+	txn.End()
+	app.ExpectSpanEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":          "OtherTransaction/Go/hello",
+				"sampled":       true,
+				"category":      "generic",
+				"nr.entryPoint": true,
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"sampled":   true,
+				"name":      "Datastore/statement/MySQL/mycollection/myoperation",
+				"category":  "datastore",
+				"component": "MySQL",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"name":      "External/example.com/all",
+				"category":  "http",
+				"component": "http",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+	})
+}
+
+func TestSpanEventAttributesSpecificallyExcluded(t *testing.T) {
+	// Test that SpanEvents.Attributes.Exclude excludes span attributes.
+	replyfn := func(reply *internal.ConnectReply) {
+		reply.AdaptiveSampler = internal.SampleEverything{}
+	}
+	cfgfn := func(cfg *Config) {
+		cfg.DistributedTracer.Enabled = true
+		cfg.CrossApplicationTracer.Enabled = false
+		cfg.SpanEvents.Attributes.Exclude = []string{
+			SpanAttributeDBStatement,
+			SpanAttributeDBInstance,
+			SpanAttributePeerAddress,
+			SpanAttributePeerHostname,
+			SpanAttributeHTTPURL,
+			SpanAttributeHTTPMethod,
+		}
+	}
+	app := testApp(replyfn, cfgfn, t)
+	txn := app.StartTransaction("hello", nil, nil)
+	segment := DatastoreSegment{
+		StartTime:          StartSegmentNow(txn),
+		Product:            DatastoreMySQL,
+		Collection:         "mycollection",
+		Operation:          "myoperation",
+		ParameterizedQuery: "myquery",
+		Host:               "myhost",
+		PortPathOrID:       "myport",
+		DatabaseName:       "dbname",
+	}
+	segment.End()
+	req, _ := http.NewRequest("GET", "http://example.com?ignore=me", nil)
+	s := StartExternalSegment(txn, req)
+	s.End()
+	txn.End()
+	app.ExpectSpanEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":          "OtherTransaction/Go/hello",
+				"sampled":       true,
+				"category":      "generic",
+				"nr.entryPoint": true,
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"sampled":   true,
+				"name":      "Datastore/statement/MySQL/mycollection/myoperation",
+				"category":  "datastore",
+				"component": "MySQL",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"name":      "External/example.com/all",
+				"category":  "http",
+				"component": "http",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+	})
+}
+
+func TestSpanEventAttributesExcluded(t *testing.T) {
+	// Test that Attributes.Exclude excludes span attributes.
+	replyfn := func(reply *internal.ConnectReply) {
+		reply.AdaptiveSampler = internal.SampleEverything{}
+	}
+	cfgfn := func(cfg *Config) {
+		cfg.DistributedTracer.Enabled = true
+		cfg.CrossApplicationTracer.Enabled = false
+		cfg.Attributes.Exclude = []string{
+			SpanAttributeDBStatement,
+			SpanAttributeDBInstance,
+			SpanAttributePeerAddress,
+			SpanAttributePeerHostname,
+			SpanAttributeHTTPURL,
+			SpanAttributeHTTPMethod,
+		}
+	}
+	app := testApp(replyfn, cfgfn, t)
+	txn := app.StartTransaction("hello", nil, nil)
+	segment := DatastoreSegment{
+		StartTime:          StartSegmentNow(txn),
+		Product:            DatastoreMySQL,
+		Collection:         "mycollection",
+		Operation:          "myoperation",
+		ParameterizedQuery: "myquery",
+		Host:               "myhost",
+		PortPathOrID:       "myport",
+		DatabaseName:       "dbname",
+	}
+	segment.End()
+	req, _ := http.NewRequest("GET", "http://example.com?ignore=me", nil)
+	s := StartExternalSegment(txn, req)
+	s.End()
+	txn.End()
+	app.ExpectSpanEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":          "OtherTransaction/Go/hello",
+				"sampled":       true,
+				"category":      "generic",
+				"nr.entryPoint": true,
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"sampled":   true,
+				"name":      "Datastore/statement/MySQL/mycollection/myoperation",
+				"category":  "datastore",
+				"component": "MySQL",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"name":      "External/example.com/all",
+				"category":  "http",
+				"component": "http",
+				"span.kind": "client",
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+	})
+}
+
+func TestSpanEventAttributesLASP(t *testing.T) {
+	// Test that security policies prevent the capture of the input query
+	// statement.
+	replyfn := func(reply *internal.ConnectReply) {
+		reply.AdaptiveSampler = internal.SampleEverything{}
+		reply.SecurityPolicies.RecordSQL.SetEnabled(false)
+	}
+	cfgfn := func(cfg *Config) {
+		cfg.DistributedTracer.Enabled = true
+		cfg.CrossApplicationTracer.Enabled = false
+	}
+	app := testApp(replyfn, cfgfn, t)
+	txn := app.StartTransaction("hello", nil, nil)
+	segment := DatastoreSegment{
+		StartTime:          StartSegmentNow(txn),
+		Product:            DatastoreMySQL,
+		Collection:         "mycollection",
+		Operation:          "myoperation",
+		ParameterizedQuery: "myquery",
+		Host:               "myhost",
+		PortPathOrID:       "myport",
+		DatabaseName:       "dbname",
+	}
+	segment.End()
+	req, _ := http.NewRequest("GET", "http://example.com?ignore=me", nil)
+	s := StartExternalSegment(txn, req)
+	s.End()
+	txn.End()
+	app.ExpectSpanEvents(t, []internal.WantEvent{
+		{
+			Intrinsics: map[string]interface{}{
+				"name":          "OtherTransaction/Go/hello",
+				"sampled":       true,
+				"category":      "generic",
+				"nr.entryPoint": true,
+			},
+			UserAttributes:  map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"sampled":   true,
+				"name":      "Datastore/statement/MySQL/mycollection/myoperation",
+				"category":  "datastore",
+				"component": "MySQL",
+				"span.kind": "client",
+			},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"db.instance":   "dbname",
+				"peer.address":  "myhost:myport",
+				"peer.hostname": "myhost",
+				"db.statement":  "'myoperation' on 'mycollection' using 'MySQL'",
+			},
+		},
+		{
+			Intrinsics: map[string]interface{}{
+				"parentId":  internal.MatchAnything,
+				"name":      "External/example.com/all",
+				"category":  "http",
+				"component": "http",
+				"span.kind": "client",
+			},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"http.url":    "http://example.com",
+				"http.method": "GET",
+			},
+		},
+	})
 }